added rocket reservation
