+++ src/components/navLink/NavLink.jsx
@@ -14,3 +14,17 @@ const navLink = ({ content, path }) => {
 }
 
 export default navLink
+
+//  {rocketsdata.data.map((e) => (
+//             <>
+//               <li key={e.id}>
+//                 {e.flickr_images && <img src={e.flickr_images} alt={e.name} />}
+//                 <h2>{e.name}</h2>
+//                 <p>
+//                   {!e.reserved && <p style={{ background: "red", padding: "40px" }}>Reserved</p>} {e.description}
+//                 </p>
+//               </li>
+//               <hr />
+//             </>
+
+//           ))}
